web-animations/animation-node/animation-node-before.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationNode before() method tests</title>
<meta name="assert" content="Inserts nodes before this animation node.">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationNode-before-void-AnimationNode...-nodes">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#dfn-insert-children">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#dfn-remove-an-animation-node">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
// Test step 1. If there is no parent animation group, terminate these steps.
test(function() {
    var node = newAnimation(createDiv(this));
    try {
        node.before(null);
    } catch(e) {
        assert_unreached('Caught unexpected exception: ' + e);
    }
}, 'AnimationNode.before() does nothing if the node has no parent animation group. ' +
    'HierarchyRequestError is not thrown in call node.before(null). ' +
    'Test Animation');

test(function() {
    var node = new AnimationGroup([]);
    try {
        node.before(null);
    } catch(e) {
        assert_unreached('Caught unexpected exception: ' + e);
    }
}, 'AnimationNode.before() does nothing if the node has no parent animation group. ' +
    'HierarchyRequestError is not thrown in call node.before(null). ' +
    'Test AnimationGroup');

test(function() {
    var node = newAnimation(createDiv(this));
    try {
        node.before(node);
    } catch(e) {
        assert_unreached('Caught unexpected exception: ' + e);
    }
}, 'AnimationNode.before() does nothing if the node has no parent animation group. ' +
    'HierarchyRequestError is not thrown if the node is inserted before itself. ' +
    'Test Animation');

test(function() {
    var node = new AnimationGroup([]);
    try {
        node.before(node);
    } catch(e) {
        assert_unreached('Caught unexpected exception: ' + e);
    }
}, 'AnimationNode.before() does nothing if the node has no parent animation group. ' +
    'HierarchyRequestError is not thrown if the node is inserted before itself. ' +
    'Test AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));

    node1.before(node2);

    assert_equals(node1.previousSibling, null, 'AnimationNode.before() should do nothing ' +
        'if the node has no parent animation group');
    assert_equals(node2.nextSibling, null, 'AnimationNode.before() should do nothing ' +
        'if the node has no parent animation group');
}, 'AnimationNode.before() does nothing if the node has no parent animation group. ' +
    'Test Animation');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = new AnimationGroup([]);

    node1.before(node2);

    assert_equals(node1.previousSibling, null, 'AnimationNode.before() should do nothing ' +
        'if the node has no parent animation group');
    assert_equals(node2.nextSibling, null, 'AnimationNode.before() should do nothing ' +
        'if the node has no parent animation group');
}, 'AnimationNode.before() does nothing if the node has no parent animation group. ' +
    'Test AnimationGroup');

// Test step 2. If any of the animation nodes in nodes is an inclusive ancestor of this animation node throw a HierarchyRequestError exception and terminate these steps.
test(function() {
    var node = newAnimation(createDiv(this));
    var parent = new AnimationGroup([node]);

    assert_throws('HierarchyRequestError', function() {
        node.before(node);
    }, 'HierarchyRequestError should be thrown if inserting the node ' +
        'before itself');
}, 'Test HierarchyRequestError is thrown if the node is inserted before itself');

test(function() {
    var node = new AnimationGroup([]);
    var parent = new AnimationGroup([node]);

    assert_throws('HierarchyRequestError', function() {
        node.before(node);
    }, 'HierarchyRequestError should be thrown if inserting the node ' +
        'before itself');
}, 'HierarchyRequestError is thrown if the node is inserted before itself. Test AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = new AnimationGroup([node1]);

    assert_throws('HierarchyRequestError', function() {
        node1.before(node2);
    }, 'HierarchyRequestError should be thrown if inserting ' +
        'node\'s parent');
}, 'HierarchyRequestError is thrown if an inclusive ancestor is inserted before the node');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = new AnimationGroup([node1]);

    assert_throws('HierarchyRequestError', function() {
        node1.before(node2);
    }, 'HierarchyRequestError should be thrown if inserting ' +
        'node\'s parent');
}, 'HierarchyRequestError is thrown if an inclusive ancestor is inserted before the node. ' +
    'Test AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = new AnimationGroup([node1]);
    var node3 = new AnimationGroup([node2]);
    var node4 = new AnimationGroup([node3]);

    assert_throws('HierarchyRequestError', function() {
        node1.before(node3);
    }, 'HierarchyRequestError should be thrown if inserting ' +
        'an inclusive ancestor');
}, 'HierarchyRequestError is thrown if an inclusive ancestor is inserted before the node. ' +
    'Inclusive ancestor is not a direct parent of the node');


test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = new AnimationGroup([node1]);
    var node3 = new AnimationGroup([node2]);
    var node4 = new AnimationGroup([node3]);

    assert_throws('HierarchyRequestError', function() {
        node1.before(node3);
    }, 'HierarchyRequestError should be thrown if inserting ' +
        'an inclusive ancestor');
}, 'HierarchyRequestError is thrown if an inclusive ancestor is inserted before the node. ' +
    'Inclusive ancestor is not a direct parent of the node for AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = newAnimation(createDiv(this));
    var node4 = new AnimationGroup([node1, node2]);

    assert_throws('HierarchyRequestError', function() {
        node1.before(node3, node4);
    }, 'HierarchyRequestError should be thrown if inserting ' +
        'an inclusive ancestor');
}, 'HierarchyRequestError is thrown if an inclusive ancestor is inserted before the node. ' +
    'Test several arguments in before() call');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = newAnimation(createDiv(this));
    var node3 = newAnimation(createDiv(this));
    var node4 = new AnimationGroup([node1, node2]);

    assert_throws('HierarchyRequestError', function() {
        node1.before(node3, node4);
    }, 'HierarchyRequestError should be thrown if inserting ' +
        'an inclusive ancestor');
}, 'HierarchyRequestError is thrown if an inclusive ancestor is inserted before the node. ' +
    'Test several arguments in before() call for AnimationGroup');

// Test step 3. Insert nodes before this animation node.
test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1]);

    node1.before(node2);

    assert_equals(node1.previousSibling, node2, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node2.parent, animationGroup, 'Node should be inserted into the tree');
    assert_equals(node2.nextSibling, node1, 'Node should be inserted into the tree ' +
        'before this node');
}, 'AnimationNode.before() inserts nodes before this node. Test Animation');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = new AnimationGroup([]);
    var animationGroup = new AnimationGroup([node1]);

    node1.before(node2);

    assert_equals(node1.previousSibling, node2, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node2.parent, animationGroup, 'Node should be inserted into the tree');
    assert_equals(node2.nextSibling, node1, 'Node should be inserted into the tree ' +
        'before this node');
}, 'AnimationNode.before() inserts nodes before this node. Test AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1, node2]);

    node1.before(node2);

    assert_equals(node1.previousSibling, node2, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node2.nextSibling, node1, 'Node should be inserted into the tree ' +
        'before this node');
}, 'AnimationNode.before() inserts nodes before this node. Inserted node is on the same ' +
    'level in the tree. Test Animation');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = new AnimationGroup([]);
    var animationGroup = new AnimationGroup([node1, node2]);

    node1.before(node2);

    assert_equals(node1.previousSibling, node2, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node2.nextSibling, node1, 'Node should be inserted into the tree ' +
        'before this node');
}, 'AnimationNode.before() inserts nodes before this node. Inserted node is on the same ' +
    'level in the tree. Test AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1, node2]);

    node2.before(node1);

    assert_equals(node1.nextSibling, node2, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node2.previousSibling, node1, 'Node should be inserted into the tree ' +
        'before this node');
}, 'Test AnimationNode.before() inserts node before this node even if inserted ' +
    'node is already before this one. Test Animation');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = new AnimationGroup([]);
    var animationGroup = new AnimationGroup([node1, node2]);

    node2.before(node1);

    assert_equals(node1.nextSibling, node2, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node2.previousSibling, node1, 'Node should be inserted into the tree ' +
        'before this node');
}, 'Test AnimationNode.before() inserts node before this node even if inserted ' +
    'node is already before this one. Test AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = newAnimation(createDiv(this));
    var node4 = newAnimation(createDiv(this));
    var group1 = new AnimationGroup([node1, node2]);
    var group2 = new AnimationGroup([node3, group1, node4]);

    node4.before(node1);

    assert_equals(node1.nextSibling, node4, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node1.parent, group2, 'Parent group of the inserted node should be changed');
    assert_equals(node1.previousSibling, group1, 'AnimationNode.before() should insert ' +
        'nodes before this node');

    assert_equals(node4.previousSibling, node1, 'Node should be inserted into the tree ' +
        'before this node');
    assert_equals(group1.nextSibling, node1, 'Node should be inserted into the tree ' +
        'before this node');

    assert_equals(node2.previousSibling, null, 'Inserted node should be removed from its ' +
        'previous position in the tree');
    assert_array_equals(group1.children, [node2], 'Inserted node should be removed from its ' +
        'previous position in the tree');
}, 'Test AnimationNode.before() inserts node before this node. The previous position ' +
    'of the inserted node is lower in the tree than current node');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = newAnimation(createDiv(this));
    var node4 = newAnimation(createDiv(this));
    var group1 = new AnimationGroup([node1, node2]);
    var group2 = new AnimationGroup([node3, group1, node4]);

    node2.before(node4);

    assert_equals(node4.previousSibling, node1, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node4.nextSibling, node2, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node4.parent, group1, 'Parent group of the inserted node should be changed');

    assert_equals(node2.previousSibling, node4, 'Node should be inserted into the tree ' +
        'before this node');
    assert_equals(node1.nextSibling, node4, 'Node should be inserted into the tree ' +
        'before this node');

    assert_equals(group1.nextSibling, null, 'Inserted node should be removed from its ' +
        'previous position in the tree');
    assert_array_equals(group2.children, [node3, group1], 'Inserted node should be ' +
        'removed from its previous position in the tree');
}, 'Test AnimationNode.before() inserts node before this node. The previous position ' +
    'of the inserted node is higher in the tree than current node, but not ancestor');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var node3 = newAnimation(createDiv(this));
    var node4 = newAnimation(createDiv(this));
    var group = new AnimationGroup([node1, node2]);
    // FIXME Is resulting order of nodes specified?
    node2.before(node3, node4);

    assert_equals(node1.nextSibling, node3, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node3.previousSibling, node1, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node3.nextSibling, node4, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node3.parent, group, 'Parent group of the inserted node should be changed');
    assert_equals(node4.previousSibling, node3, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node4.nextSibling, node2, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node4.parent, group, 'Parent group of the inserted node should be changed');
    assert_equals(node2.previousSibling, node4, 'AnimationNode.before() should insert ' +
        'nodes before this node');
}, 'Test AnimationNode.before() inserts several nodes before this node. Test Animation');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = new AnimationGroup([]);
    var node3 = newAnimation(createDiv(this));
    var node4 = new AnimationGroup([]);
    var group = new AnimationGroup([node1, node2]);
    // FIXME Is resulting order of nodes specified?
    node2.before(node3, node4);

    assert_equals(node1.nextSibling, node3, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node3.previousSibling, node1, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node3.nextSibling, node4, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node3.parent, group, 'Parent group of the inserted node should be changed');
    assert_equals(node4.previousSibling, node3, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node4.nextSibling, node2, 'AnimationNode.before() should insert ' +
        'nodes before this node');
    assert_equals(node4.parent, group, 'Parent group of the inserted node should be changed');
    assert_equals(node2.previousSibling, node4, 'AnimationNode.before() should insert ' +
        'nodes before this node');
}, 'Test AnimationNode.before() inserts several nodes before this node. Test AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(node2);

    assert_equals(node2.player, player, 'The node should be associated with it own player');
    node1.before(node2);
    assert_equals(node2.player, null, 'The node should be disassociated from its player');
}, 'Test AnimationNode.before() disassociates the inserted node from the player, ' +
    'if node is directly associated with a player. Test Animation');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = new AnimationGroup([]);
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(node2);

    assert_equals(node2.player, player, 'The node should be associated with it own player');
    node1.before(node2);
    assert_equals(node2.player, null, 'The node should be disassociated from its player');

}, 'Test AnimationNode.before() disassociates the inserted node from the player, ' +
    'if node is directly associated with a player. Test AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1]);
    var animationGroupPlayer = document.timeline.play(animationGroup);

    assert_equals(node2.player, null, 'The node should not be associated with it own player');
    node1.before(node2);
    assert_equals(node2.player, animationGroupPlayer,
        'The node should be associated with animation group player');
}, 'Test AnimationNode.before() associates the inserted node with animation group player. ' +
    'Test Animation');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = new AnimationGroup([]);
    var animationGroup = new AnimationGroup([node1]);
    var animationGroupPlayer = document.timeline.play(animationGroup);

    assert_equals(node2.player, null, 'The node should not be associated with it own player');
    node1.before(node2);
    assert_equals(node2.player, animationGroupPlayer,
        'The node should be associated with animation group player');
}, 'Test AnimationNode.before() associates the inserted node with animation group player. ' +
    'Test AnimationGroup');

test(function() {
    var node1 = newAnimation(createDiv(this));
    var node2 = newAnimation(createDiv(this));
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(node2);
    var animationGroupPlayer = document.timeline.play(animationGroup);

    assert_equals(node2.player, player, 'The node should be associated with it own player');
    node1.before(node2);
    assert_equals(node2.player, animationGroupPlayer,
        'The node should be associated with animation group player');
}, 'Test AnimationNode.before() associates the inserted node with animation group player, ' +
    'if the node is directly associated with a player. Test Animation');

test(function() {
    var node1 = new AnimationGroup([]);
    var node2 = new AnimationGroup([]);
    var animationGroup = new AnimationGroup([node1]);
    var player = document.timeline.play(node2);
    var animationGroupPlayer = document.timeline.play(animationGroup);

    assert_equals(node2.player, player, 'The node should be associated with it own player');
    node1.before(node2);
    assert_equals(node2.player, animationGroupPlayer,
        'The node should be associated with animation group player');
}, 'Test AnimationNode.before() associates the inserted node with animation group player, ' +
    'if the node is directly associated with a player. Test AnimationGroup');
</script>
</body>
